Commit update permisos

+++ app/Http/Controllers/PermissionController.php
@@ -10,6 +10,7 @@ class PermissionController extends Controller
 {
     public function index(Request $request)
     {
+        abort_if(Gate::denies('Listado_permisos'),redirect()->route('madre')->with('error','No tiene acceso'));
         $busqueda = trim($request->get('busqueda'));
 
         $permisos = Permission::orderby('permissions.id','DESC')
@@ -24,12 +25,13 @@ class PermissionController extends Controller
 
     public function create()
     {
+        abort_if(Gate::denies('Nuevo_permiso'),redirect()->route('madre')->with('error','No tiene acceso'));
         return view('permisos/crearpermiso');
     }
 
     public function store(Request $request, $perm=-1)
     {
-
+        abort_if(Gate::denies('Nuevo_permiso'),redirect()->route('madre')->with('error','No tiene acceso'));
         $rules=[
             'name' => 'required|max:31|min:5|unique:permissions,name',
             'descripcion'=> 'required|max:83',
@@ -63,6 +65,7 @@ class PermissionController extends Controller
     //FUNCIÓN PARA EDITAR LOS PERMISOS
     public function editar($id)
     {
+        abort_if(Gate::denies('Editar_permisos'),redirect()->route('madre')->with('error','No tiene acceso'));
         $permisos = Permission::findOrFail($id);
         return view('permisos/permisoEditar')
             ->with('permisos', $permisos);
@@ -71,6 +74,7 @@ class PermissionController extends Controller
 
       //Función para guardar los datos actualizados AL EDITAR UN PERMISO
       public function update(Request $request, $id){
+        abort_if(Gate::denies('Editar_permisos'),redirect()->route('madre')->with('error','No tiene acceso'));
         //Validar campos del formulario editar
         $rules= [
             'name' => 'required|max:31|min:5|unique:permissions,name,'.$id,
@@ -107,6 +111,7 @@ class PermissionController extends Controller
     }
 
     public function destroy($id){
+        abort_if(Gate::denies('Eliminar_permisos'),redirect()->route('madre')->with('error','No tiene acceso'));
         Permission::destroy($id);
 
         return redirect()->route('permisos.lista')
